requirements.txt
Django==1.6
gunicorn==18.0
six==1.4.1
django-braces==1.2.2
#MySQL-python==1.2.4
South==0.8.3
cssmin==0.2.0
django-assets==0.8
django-widget-tweaks==1.3
jsmin==2.0.8
pytz==2013.8
requests==2.0.1
webassets==0.8
wsgiref==0.1.2
supervisor==3.0
Django-Select2==4.2.2
psycopg2==2.5.1
raven==3.5.2
